pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "Team_5_AI_Tutor"
version = "0.0.1"
description = "Хакатон за третий семестр. Команда 5. Задача 1. Разработка AI-репетитора по техническим дисциплинам, Cloud.ru"
authors = [
  { name = "Your name (or your organization/company/team)" },
]

readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    
]
dependencies = [
    "accelerate>=1.12.0",
    "jupyter>=1.1.1",
    "loguru",
    "mkdocs",
    "pdf2image>=1.17.0",
    "pillow>=12.0.0",
    "pip",
    "pytest",
    "python-dotenv",
    "ruff",
    "torch>=2.9.1",
    "tqdm",
    "transformers>=4.57.3",
    "typer",
]
requires-python = "~=3.12.0"


[tool.ruff]
line-length = 99
src = ["Team_5_AI_Tutor"]
include = ["pyproject.toml", "Team_5_AI_Tutor/**/*.py"]

[tool.ruff.lint]
extend-select = ["I"]  # Add import sorting

[tool.ruff.lint.isort]
known-first-party = ["Team_5_AI_Tutor"]
force-sort-within-sections = true

[tool.uv.sources]
transformers = { git = "https://github.com/huggingface/transformers", rev = "82a06db03535c49aa987719ed0746a76093b1ec4" }

